Improved the complexity of output filtering (Go).
I have never liked my output filtering function in
the go version of the code.  It is much cleaner in
Scala and F#, so I went into the Go version and
made an 'unfold'-type operation by hand. It
goes from []token to (string,[]token), and is
called repeatedly until the []token returned
is empty.  The result is much nicer than
the state machine I had used before.