Made the initial jump to Autotools/Automake.
I wanted to use this small project as a test, as it has been
years since the last time I created a GNU-like build setup.
It's a behemoth, but you _do_ get some benefits for your
trouble, like: support for out-of-tree builds, distribution
packaging, some semblance of portability, a familiar user
experience (configure&&make).  So, I will try to stick with it.

I used the `foreign` option to automake so it doesn't insist
that I check in a changelog.