UWP version tries to handle big-endian, starting from explorer.
I do a basic check for endianness, and byte-swap the floats on big-endian. Note:
this will only work on machines that are all-big or all-little, which I think is almost
all modern machines.

Also, I added a basic check to try to see if someone started the program via the
start menu, and if so I make them press a key before exiting the program. This way,
the user gets to actually read the error message.  You can fool it with: cls; bascat  but
that is acceptable to me.