Reorganized the module and library.
I've realized how much nicer it is to write the cmdlet in powershell, and just depend
on a standard dotnet dll (with no Automation dll dependencies!) to do the work.  It
really is a best-of-both-worlds solution, because I get a c# library I can use in a
c# program, and also interactive use via powershell cmdlets on demand.

I will be writing many more library+cmdlet-wrapper programs in the future, I think.